--- src/helpers/totalTvsAvailable.js
@@ -1,5 +1,0 @@
-function getTotalTvsAvailable(getTvsSold, getTvsPurchased) {
-    return getTvsPurchased - getTvsSold;
-}
-
-export default getTotalTvsAvailable;
