--- src/helpers/tvNameGenerator.js
@@ -1,5 +1,0 @@
-function generateTvName(bestSellingTv) {
-    return `${bestSellingTv.brand} ${bestSellingTv.type} - ${bestSellingTv.name}`;
-}
-
-export {generateTvName};
