--- src/helpers/tvPriceGenerator.js
@@ -1,5 +1,0 @@
-function generateTvPrice(bestSellingTv) {
-    return `€${bestSellingTv.price},-`;
-}
-
-export {generateTvPrice};
