--- src/helpers/tvScreenSizeGenerator.js
@@ -1,8 +1,0 @@
-import {inchToCm} from "./inchToCm.js";
-
-function generateTvScreenSize(bestSellingTv) {
-    return bestSellingTv.availableSizes.map(size =>
-        `${size} inch (${inchToCm(size)} cm)`).join(' | ');
-}
-
-export {generateTvScreenSize};
